Workaround libiconv URL Timeout Read
Apparently the link we're using is a redirect, but to the same URL with
a / at the end. That doesn't seem to be the problem though. Using curl
to see what's happening, the link is resolving, it's just ftp.gnu.org is
taking a time long to respond. Ignore read timeouts in general, also saw
one on Strawberry Perl website while testing.