Merged PR 15071: Use an IDL Map for compact representation of RtpsRelay statistics
IDL Map data types are a new feature in OpenDDS.  Using a Map allows the JSON representation of the statistics, which appears in the log, to be much more compact.

This PR merges the supporting code for Maps (OpenDDS 4999) along with the change that uses a Map type for RtpsRelay Statistics (OpenDDS 5024).

Related work items: #427230

 Conflicts:
	tools/dds/rtpsrelaylib/Relay.idl
	tools/rtpsrelay/RelayStatisticsReporter.cpp

 It looks like you may be committing a cherry-pick.
 If this is not correct, please run
	git update-ref -d CHERRY_PICK_HEAD
 and try again.